requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.128.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.21
sqladmin==0.22.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
WTForms==3.1.2
httpx
python-dotenv
sendgrid python-http-clien







